Content Model: improve formatWithContentModel 1 (#2001)
* Content Model: Improve cache behavior

* fix build

* Content Model: improve formatWithContentModel
